requirements.txt
# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-asyncio>=0.18.0

# Linting
flake8>=4.0.0
black>=22.0.0
isort>=5.10.0
mypy>=0.910

fastapi>=0.68.0
geojson>=2.5.0
shapely>=1.8.0
setuptools==76.0.0

# CDK
boto3==1.35.88
aws-cdk-lib>=2.0.0
constructs>=10.0.0
redis>=4.3.0

